datos-xm-simen-starter/scripts/requirements.txt
pandas>=2.1.0
pyarrow>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
psycopg2-binary>=2.9.9
timescale-vector~=0.0.5 ; python_version>='3.10' # harmless if not used
clickhouse-driver>=0.2.9
pymongo>=4.6.0
boto3>=1.28.0
# Optional: library for XM data (uncomment if you want to try it)
# pydataxm>=0.4.0
